backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-swagger-ui==4.11.1
flask-sqlalchemy==3.1.1
PyPDF2==3.0.1
python-dotenv==1.0.0
requests==2.31.0
flask-jwt-extended==4.5.3
werkzeug==3.0.1
flasgger==0.9.7.1
minio==7.2.0
pytest==8.1.1
pytest-flask==1.3.0
pytest-cov==5.0.0
pytest-mock==3.14.0
